+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/auto_classes/auto1.java
@@ -37,48 +37,48 @@ public class auto1 extends OpMode {
 
         if (position == 1) {
             //starting facing wall, and 42.25'' away from the other wall
-            Robot.DCG.driveToPosition(0, -65.75, 0);
-            Robot.DCG.driveToPosition(20, 0, 0);
-            Robot.DCG.driveToPosition(0, 47.5, 0);
-            Robot.DCG.driveToPosition(0, -47.5, 0);
-            Robot.DCG.driveToPosition(-46.75, 0, 0);
-            Robot.DCG.driveToPosition(0, -25, 0);
+            Robot.DCGm.driveToPosition(0, -65.75, 0);
+            Robot.DCGm.driveToPosition(20, 0, 0);
+            Robot.DCGm.driveToPosition(0, 47.5, 0);
+            Robot.DCGm.driveToPosition(0, -47.5, 0);
+            Robot.DCGm.driveToPosition(-46.75, 0, 0);
+            Robot.DCGm.driveToPosition(0, -25, 0);
 
             while (Robot.CSG.FrontColorSensor.argb() < averageColorFront) {
-                Robot.DCG.driveToPosition(0, -65.75, 0);
-                Robot.DCG.driveToPosition(20, 0, 0);
-                Robot.DCG.driveToPosition(0, 47.5, 0);
-                Robot.DCG.driveToPosition(0, -47.5, 0);
-                Robot.DCG.driveToPosition(-46.75, 0, 0);
-                Robot.DCG.driveToPosition(0, -25, 0);
+                Robot.DCGm.driveToPosition(0, -65.75, 0);
+                Robot.DCGm.driveToPosition(20, 0, 0);
+                Robot.DCGm.driveToPosition(0, 47.5, 0);
+                Robot.DCGm.driveToPosition(0, -47.5, 0);
+                Robot.DCGm.driveToPosition(-46.75, 0, 0);
+                Robot.DCGm.driveToPosition(0, -25, 0);
             }
 
             while (Robot.CSG.FrontColorSensor.argb() < averageColorFront) {
-                Robot.DCG.driveToPosition(-1, 0, 0);
+                Robot.DCGm.driveToPosition(-1, 0, 0);
                 offset++;
             }
 
             // at this point, the robot will pick up the skystone
-            Robot.DCG.driveToPosition(offset,0,0);
-            Robot.DCG.driveToPosition(0,-43,0);
-            Robot.DCG.driveToPosition(26.75,0,0);
-            Robot.DCG.driveToPosition(0,68,0);
-            Robot.DCG.driveToPosition(20,0,0);
-            Robot.DCG.driveToPosition(0,0,180);
+            Robot.DCGm.driveToPosition(offset,0,0);
+            Robot.DCGm.driveToPosition(0,-43,0);
+            Robot.DCGm.driveToPosition(26.75,0,0);
+            Robot.DCGm.driveToPosition(0,68,0);
+            Robot.DCGm.driveToPosition(20,0,0);
+            Robot.DCGm.driveToPosition(0,0,180);
 
             //At this point, the robot will place the stone in the foundation.
-            Robot.DCG.driveToPosition(0,0,180);
-            Robot.DCG.driveToPosition(-20,0,0);
-            Robot.DCG.driveToPosition(0,-68,0);
+            Robot.DCGm.driveToPosition(0,0,180);
+            Robot.DCGm.driveToPosition(-20,0,0);
+            Robot.DCGm.driveToPosition(0,-68,0);
 
             while(Robot.CSG.BottomColorSensor.argb() < avarageColorBottom) {
-                Robot.DCG.driveToPosition(0,0,180);
-                Robot.DCG.driveToPosition(-20,0,0);
-                Robot.DCG.driveToPosition(0,-68,0);
+                Robot.DCGm.driveToPosition(0,0,180);
+                Robot.DCGm.driveToPosition(-20,0,0);
+                Robot.DCGm.driveToPosition(0,-68,0);
             }
 
             while(Robot.CSG.BottomColorSensor.argb() < avarageColorBottom) {
-                Robot.DCG.driveToPosition(-1,0,0);
+                Robot.DCGm.driveToPosition(-1,0,0);
             }
 
             position = 3;
@@ -86,38 +86,38 @@ public class auto1 extends OpMode {
         if (position == 2) {
 
             //Starting at wall, facing away, and 50" away from other wall.
-            Robot.DCG.driveToPosition(0,29,0);
+            Robot.DCGm.driveToPosition(0,29,0);
 
             while (Robot.CSG.FrontColorSensor.argb() < averageColorFront) {
-                Robot.DCG.driveToPosition(0,29,0);
+                Robot.DCGm.driveToPosition(0,29,0);
             }
 
             while (Robot.CSG.FrontColorSensor.argb() < averageColorFront) {
-                Robot.DCG.driveToPosition(-1, 0, 0);
+                Robot.DCGm.driveToPosition(-1, 0, 0);
                 offset++;
             }
 
             //At this point, the robot will pick up the skystone.
-            Robot.DCG.driveToPosition(offset,0,0);
-            Robot.DCG.driveToPosition(0,-43,0);
-            Robot.DCG.driveToPosition(26.75,0,0);
-            Robot.DCG.driveToPosition(0,68,0);
-            Robot.DCG.driveToPosition(20,0,0);
-            Robot.DCG.driveToPosition(0,0,180);
+            Robot.DCGm.driveToPosition(offset,0,0);
+            Robot.DCGm.driveToPosition(0,-43,0);
+            Robot.DCGm.driveToPosition(26.75,0,0);
+            Robot.DCGm.driveToPosition(0,68,0);
+            Robot.DCGm.driveToPosition(20,0,0);
+            Robot.DCGm.driveToPosition(0,0,180);
 
             //At this point, the robot will place the stone in the foundation.
-            Robot.DCG.driveToPosition(0,0,180);
-            Robot.DCG.driveToPosition(-20,0,0);
-            Robot.DCG.driveToPosition(0,-68,0);
+            Robot.DCGm.driveToPosition(0,0,180);
+            Robot.DCGm.driveToPosition(-20,0,0);
+            Robot.DCGm.driveToPosition(0,-68,0);
 
             while(Robot.CSG.BottomColorSensor.argb() < avarageColorBottom) {
-                Robot.DCG.driveToPosition(0,0,180);
-                Robot.DCG.driveToPosition(-20,0,0);
-                Robot.DCG.driveToPosition(0,-68,0);
+                Robot.DCGm.driveToPosition(0,0,180);
+                Robot.DCGm.driveToPosition(-20,0,0);
+                Robot.DCGm.driveToPosition(0,-68,0);
             }
 
             while(Robot.CSG.BottomColorSensor.argb() < avarageColorBottom) {
-                Robot.DCG.driveToPosition(-1,0,0);
+                Robot.DCGm.driveToPosition(-1,0,0);
             }
             
             position = 3;
@@ -135,6 +135,6 @@ public class auto1 extends OpMode {
     //Initialized by: Stop / runs once
     @Override
     public void stop() {
-        Robot.DCG.setPower(new double[]{0,0,0,0});
+        Robot.DCGm.setPower(new double[]{0,0,0,0});
     }
 }
